+++ eggs/naaya.content.meeting/naaya/content/meeting/countries.py
@@ -1,42 +1,18 @@
-country_from_country_code = {
-    'nl': 'Netherlands',
-    'cy': 'Cyprus',
-    'lv': 'Latvia',
-    'gb': 'United Kingdom',
-    'es': 'Spain',
-    'at': 'Austria',
-    'de': 'Germany',
-    'fr': 'France',
-    'pl': 'Poland',
-    'lt': 'Lithuania',
-    'lu': 'Luxembourg',
-    'li': 'Liechtenstein',
-    'mt': 'Malta',
-    'hu': 'Hungary',
-    'ie': 'Ireland',
-    'sk': 'Slovakia',
-    'no': 'Norway',
-    'tr': 'Turkey',
-    'fi': 'Finland',
-    'ie': 'Ireland',
-    'is': 'Iceland',
-    'se': 'Sweden',
-    'dk': 'Denmark',
-    'bg': 'Bulgaria',
-    'it': 'Italy',
-    'ee': 'Estonia',
-    'ro': 'Romania',
-    'pt': 'Portugal',
-    'si': 'Slovenia',
-    'cz': 'Czech Republic',
-    'be': 'Belgium',
-    'gr': 'Greece',
-    'ch': 'Switzerland',
-    'hr': 'Croatia',
-    'al': 'Albania',
-    'mk': 'Macedonia, The Former Yugoslav Republic',
-    'me': 'Montenegro',
-    'rs': 'Serbia',
-    'ba': 'Bosnia and Herzegovina',
-    'xk': 'Kosovo under the UNSCR 1244/99',
-}
+from eea.ldapadmin.countries import get_country, load_countries
+
+
+def get_country_name(code):
+    """ """
+    if code is None:
+        return '-'
+    country = get_country(code)
+    if country['name']:
+        return country['name']
+    else:
+        return code
+
+
+def get_countries():
+    return sorted(
+        [(country['code'], country['name']) for country in load_countries()],
+        key=lambda x: x[1])